+++ next-app/src/app/page.tsx
@@ -1,23 +1,11 @@
 "use client";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import styles from "./page.module.scss";
 
 export default function Home() {
   const [jobId, setJobId] = useState<string | null>(null);
   const [error, setError] = useState<string | null>(null);
-  const [jobs, setJobs] = useState<{ jobId: string; state: string }[]>([]);
 
-  async function fetchJobs() {
-    const res = await fetch("/api/jobs");
-    if (res.ok) {
-      const data = await res.json();
-      setJobs(data.jobs as { jobId: string; state: string }[]);
-    }
-  }
-
-  useEffect(() => {
-    fetchJobs();
-  }, []);
 
   async function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
     e.preventDefault();
@@ -30,7 +18,6 @@ export default function Home() {
     if (res.ok) {
       const data = await res.json();
       setJobId(data.jobId as string);
-      fetchJobs();
     } else {
       setError("failed to create job");
     }
@@ -49,13 +36,6 @@ export default function Home() {
           JobId: <a href={`/jobs/${jobId}`}>{jobId}</a>
         </p>
       )}
-      <ul>
-        {jobs.map((job) => (
-          <li key={job.jobId}>
-            <a href={`/jobs/${job.jobId}`}>{job.jobId}</a> - {job.state}
-          </li>
-        ))}
-      </ul>
       {error && <p>{error}</p>}
     </div>
   );
